Split solver.hpp into several files
