+++ tests/unit_tests/restruct_poc/test_solver.cpp
@@ -4,7 +4,7 @@
 
 #include "src/restruct_poc/beams.hpp"
 #include "src/restruct_poc/beams_input.hpp"
-#include "src/restruct_poc/solver.hpp"
+#include "src/restruct_poc/Solver.hpp"
 #include "tests/unit_tests/gen_alpha_poc/test_utilities.h"
 #include "src/restruct_poc/InitializeConstraints.hpp"
 #include "src/restruct_poc/PredictNextState.hpp"
